requirements.txt
Django>=5.2
gunicorn
python-dotenv
psycopg2-binary
dj-database-url
requests
whitenoise
django-allauth
django-widget-tweaks
django-crispy-forms
